Fix: rust-test Has Been Red on This Branch, on clippy
`rust-test` runs `cargo clippy -- -D warnings`, and this branch has been failing
it since before the compat-residue work -- 23 errors in lib.rs and 4 more in the
test target. Every one is a lint, not a behaviour change:

  16  needless_borrow      `&face` / `&part` where the callee takes a reference
  10  useless_conversion   `u8::from(x)` where x is already u8, because an
                           archived u8 IS a u8
   1  collapsible_if

Applied with `cargo clippy --fix`, so the edits are the ones the lint itself
proposes. 175 tests pass unchanged.

Worth saying why it went unnoticed: the gate this work has been running is
`cargo test` plus `cargo check`, and neither runs clippy -- so every local run
was green while CI was red, and the PR's own checks were never looked at. A
gate that does not include what CI includes is not the gate.

+++ card_engine/src/tests.rs
@@ -11544,9 +11544,9 @@ fn compat_fields_survive_the_archive_round_trip() {
     assert_eq!(a.multiverse_ids.len(), 3);
 
     // Bitsets are only useful if membership survives independently of each other.
-    assert_ne!(u8::from(a.games) & GAME_PAPER, 0);
-    assert_ne!(u8::from(a.games) & GAME_ARENA, 0);
-    assert_eq!(u8::from(a.games) & 0b0000_0010, 0, "mtgo was not set");
+    assert_ne!(a.games & GAME_PAPER, 0);
+    assert_ne!(a.games & GAME_ARENA, 0);
+    assert_eq!(a.games & 0b0000_0010, 0, "mtgo was not set");
     assert_ne!(u16::from(a.flags) & COMPAT_PROMO, 0);
     assert_ne!(u16::from(a.flags) & COMPAT_REPRINT, 0);
     assert_eq!(u16::from(a.flags) & COMPAT_TEXTLESS, 0, "textless was not set");
@@ -11578,7 +11578,7 @@ fn absent_compat_values_stay_absent() {
     assert_eq!(u16::from(a.lang_id), VOCAB_NONE, "absent lang is the sentinel, not vocab id 0");
     assert_eq!(u16::from(a.set_type_id), VOCAB_NONE);
     assert_eq!(u16::from(a.set_vid), VOCAB_NONE);
-    assert_eq!(u8::from(a.games), 0);
+    assert_eq!(a.games, 0);
     assert_eq!(u16::from(a.flags), 0);
     assert!(a.multiverse_ids.is_empty());
     assert!(a.promo_types.is_empty());
